can't create events as student anymore

+++ source/Controllers/EventsController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Data;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Models;
@@ -15,10 +16,12 @@ namespace TomorrowC18ProjectOOP.Controllers
     public class EventsController : Controller
     {
         private readonly Context _context;
+        private readonly UserManager<Profile> userManager;
 
-        public EventsController(Context context)
+        public EventsController(Context context, UserManager<Profile> _userManager)
         {
             _context = context;
+            userManager = _userManager;
         }
 
         // GET: api/Events
@@ -51,34 +54,40 @@ namespace TomorrowC18ProjectOOP.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutEvent([FromRoute] int id, [FromBody] CalendarEvent @event)
         {
-            if (!ModelState.IsValid)
+            var userid = userManager.GetUserId(HttpContext.User);
+            Profile user = userManager.FindByIdAsync(userid).Result;
+            if (user.levelAccess != 1)
             {
-                return BadRequest(ModelState);
-            }
-
-            if (id != @event.Id)
-            {
-                return BadRequest();
-            }
-
-            _context.Entry(@event).State = EntityState.Modified;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!EventExists(id))
+                if (!ModelState.IsValid)
                 {
-                    return NotFound();
+                    return BadRequest(ModelState);
                 }
-                else
-                {
-                    throw;
-                }
-            }
 
+                if (id != @event.Id)
+                {
+                    return BadRequest();
+                }
+
+                _context.Entry(@event).State = EntityState.Modified;
+
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!EventExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return NoContent();
+            }
             return NoContent();
         }
 
@@ -86,36 +95,48 @@ namespace TomorrowC18ProjectOOP.Controllers
         [HttpPost]
         public async Task<IActionResult> PostEvent([FromBody] CalendarEvent @event)
         {
-            if (!ModelState.IsValid)
+            var userid = userManager.GetUserId(HttpContext.User);
+            Profile user = userManager.FindByIdAsync(userid).Result;
+            if (user.levelAccess != 1)
             {
-                return BadRequest(ModelState);
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+
+                _context.CalendarEvent.Add(@event);
+                await _context.SaveChangesAsync();
+
+                return CreatedAtAction("GetEvent", new { id = @event.Id }, @event);
             }
-
-            _context.CalendarEvent.Add(@event);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction("GetEvent", new { id = @event.Id }, @event);
+            return NoContent();
         }
 
         // DELETE: api/Events/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteEvent([FromRoute] int id)
         {
-            if (!ModelState.IsValid)
+            var userid = userManager.GetUserId(HttpContext.User);
+            Profile user = userManager.FindByIdAsync(userid).Result;
+            if (user.levelAccess != 1)
             {
-                return BadRequest(ModelState);
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
+
+                var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
+                if (@event == null)
+                {
+                    return NotFound();
+                }
+
+                _context.CalendarEvent.Remove(@event);
+                await _context.SaveChangesAsync();
+
+                return Ok(@event);
             }
-
-            var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
-
-            _context.CalendarEvent.Remove(@event);
-            await _context.SaveChangesAsync();
-
-            return Ok(@event);
+            return NoContent();
         }
 
         private bool EventExists(int id)
@@ -127,36 +148,42 @@ namespace TomorrowC18ProjectOOP.Controllers
         [HttpPut("{id}/move")]
         public async Task<IActionResult> MoveEvent([FromRoute] int id, [FromBody] EventMoveParams param)
         {
-            if (!ModelState.IsValid)
+            var userid = userManager.GetUserId(HttpContext.User);
+            Profile user = userManager.FindByIdAsync(userid).Result;
+            if (user.levelAccess != 1)
             {
-                return BadRequest(ModelState);
-            }
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
 
-            var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
-
-            @event.Start = param.Start;
-            @event.End = param.End;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!EventExists(id))
+                var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
+                if (@event == null)
                 {
                     return NotFound();
                 }
-                else
-                {
-                    throw;
-                }
-            }
 
+                @event.Start = param.Start;
+                @event.End = param.End;
+
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!EventExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return NoContent();
+            }
             return NoContent();
         }
 
@@ -164,38 +191,43 @@ namespace TomorrowC18ProjectOOP.Controllers
         [HttpPut("{id}/color")]
         public async Task<IActionResult> SetEventColor([FromRoute] int id, [FromBody] EventColorParams param)
         {
-            if (!ModelState.IsValid)
+            var userid = userManager.GetUserId(HttpContext.User);
+            Profile user = userManager.FindByIdAsync(userid).Result;
+            if (user.levelAccess != 1)
             {
-                return BadRequest(ModelState);
-            }
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest(ModelState);
+                }
 
-            var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
-            if (@event == null)
-            {
-                return NotFound();
-            }
-
-            @event.Color = param.Color;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!EventExists(id))
+                var @event = await _context.CalendarEvent.SingleOrDefaultAsync(m => m.Id == id);
+                if (@event == null)
                 {
                     return NotFound();
                 }
-                else
-                {
-                    throw;
-                }
-            }
 
+                @event.Color = param.Color;
+
+                try
+                {
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!EventExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+
+                return NoContent();
+            }
             return NoContent();
         }
-
     }
 
     public class EventMoveParams
